New Independent Resolution Method
The new method to rescale the screen now draws the scene onto a
RenderTarget2D at the Virtual Coordinates, and then renders it at the
scale of the window with proper resizing to the centre of the window.

This means that there is now a new RGlobal property called RenderBuffer
which will always store the render of the previous frame. It also means
that the Resolution class has had a major reduction, removing the matrix
calculation classes.